+++ lib/fastpix_client/models/components/createplaybackid_domains.rb
@@ -19,7 +19,7 @@ module FastpixClient
         # A list of domains that are explicitly blocked from accessing the resource.
         field :deny, Crystalline::Nilable.new(Crystalline::Array.new(::String)), { 'format_json': { 'letter_case': ::FastpixClient::Utils.field_name('deny') } }
 
-        sig { params(default_policy: T.nilable(Models::Components::PolicyAction), allow: T.nilable(T::Array[::String]), deny: T.nilable(T::Array[::String])).void }
+        sig { params(default_policy: T.nilable(T.any(Models::Components::PolicyAction, String)), allow: T.nilable(T::Array[::String]), deny: T.nilable(T::Array[::String])).void }
         def initialize(default_policy: nil, allow: nil, deny: nil)
           @default_policy = default_policy
           @allow = allow