+++ lib/fastpix_client/models/operations/list_playback_ids_data.rb
@@ -19,7 +19,7 @@ module FastpixClient
         # Restrictions applied to this playback ID.
         field :access_restrictions, Crystalline::Nilable.new(Models::Operations::ListPlaybackIdsAccessRestrictions), { 'format_json': { 'letter_case': ::FastpixClient::Utils.field_name('accessRestrictions') } }
 
-        sig { params(id: T.nilable(::String), access_policy: T.nilable(Models::Operations::ListPlaybackIdsAccessPolicy), access_restrictions: T.nilable(Models::Operations::ListPlaybackIdsAccessRestrictions)).void }
+        sig { params(id: T.nilable(::String), access_policy: T.nilable(T.any(Models::Operations::ListPlaybackIdsAccessPolicy, String)), access_restrictions: T.nilable(Models::Operations::ListPlaybackIdsAccessRestrictions)).void }
         def initialize(id: nil, access_policy: nil, access_restrictions: nil)
           @id = id
           @access_policy = access_policy